pyproject.toml
[project]
name = "mlx-grpc-inference"
version = "0.1.0"
description = "Distributed MLX inference system using gRPC"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mlx>=0.21.0",
    "mlx-lm>=0.20.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.2",
    "rich>=13.9.0",
    "psutil>=6.1.0",
    "aiohttp>=3.11.0",
    "numpy>=1.26.0",
    "lz4>=4.4.4",
    "zstandard>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-xdist>=3.6.0",
    "pytest-timeout>=2.3.1",
    "black>=25.1.0",
    "ruff>=0.12.7",
    "mypy>=1.13.0",
    "pre-commit>=4.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "pytest-timeout>=2.4.0",
    "pytest-xdist>=3.8.0",
]
